🔨 refactor(VSecM Sentinel): refactor RunInitCommand
Split up the file into smaller logical modules for better readability
and management.
