+++ app/sentinel/background/initialization/io_test.go
@@ -8,22 +8,4 @@
 >/'  SPDX-License-Identifier: BSD-2-Clause
 */
 
-package main
-
-import (
-	"fmt"
-	"log"
-	"os"
-
-	"github.com/vmware-tanzu/secrets-manager/core/system"
-)
-
-func main() {
-	log.Println(
-		"VSecM Keystone",
-		fmt.Sprint("v%s", os.Getenv("APP_VERSION")),
-	)
-
-	// Run on the main thread to wait forever.
-	system.KeepAlive()
-}
+package initialization
